Refatorando nomes de classes Alterando método de busca de usuário por nome de usuário ou e-mail
